Implemented Relief

+++ src/Utilities.java
@@ -1,11 +1,11 @@
 import java.util.ArrayList;
 
 public class Utilities {
-	public static int[] targetCounter(int targetDegree, ArrayList<Record> records, int targetAttributeIndex,
+	public static Double[] targetCounter(int targetDegree, ArrayList<Record> records, int targetAttributeIndex,
 			ArrayList<Attribute> attributes) {
-		int[] countKeeper = new int[targetDegree];
+		Double[] countKeeper = new Double[targetDegree];
 		for (int t = 0; t < countKeeper.length; t++) {
-			countKeeper[t] = 0;
+			countKeeper[t] = 0.0;
 		}
 		for (Record rec : records) {
 			for (int i = 0; i < targetDegree; i++) {
